+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Web-IO: on error (e. g. device not reachable) - retry after 10 s latest
+
 ## 2.0.0
 
 ### Added